fix(session-rescue): preserve exact cache-stable recovery
