+++ docs/public-evidence.md
@@ -45,11 +45,11 @@ These results measure artifact integrity and recovery, **not answer quality** or
 
 ### Recovery-resilience holdout
 
-The committed v4 recovery-resilience revalidation records **66/66** exact entries for Entroly and **66/66** for the Headroom 0.31.0 comparison under the frozen holdout protocol. This result establishes **parity, not leadership** for that recovery-integrity workload.
+The committed v5 recovery-resilience revalidation records **66/66** exact entries for Entroly and **66/66** for the Headroom 0.31.0 comparison under the frozen holdout protocol. This result establishes **parity, not leadership** for that recovery-integrity workload.
 
 The clean revalidation reproduced no Headroom worker errors. It therefore does not permit a public leadership claim and **does not establish universal recovery superiority**, production reliability, task-quality improvement, or provider-cost reduction.
 
-- [Artifact](../benchmarks/results/recovery_resilience_holdout_revalidation_v4.json)
+- [Artifact](../benchmarks/results/recovery_resilience_holdout_revalidation_v5.json)
 - [Protocol implementation](../benchmarks/recovery_resilience.py)
 
 ### WITNESS HaluEval-QA
